Fixed signed/unsigned warnings

+++ src/alignment/Matcher.cpp
@@ -6,8 +6,8 @@ Matcher::Matcher(BaseMatrix* m, int maxSeqLen){
     this->m = m;
     this->maxSeqLen = maxSeqLen;
     this->tinySubMat = new int8_t[m->alphabetSize*m->alphabetSize];
-    for (size_t i = 0; i < m->alphabetSize; i++) {
-        for (size_t j = 0; j < m->alphabetSize; j++) {
+    for (int i = 0; i < m->alphabetSize; i++) {
+        for (int j = 0; j < m->alphabetSize; j++) {
             tinySubMat[i*m->alphabetSize + j] = m->subMatrix[i][j];
         }
     }
@@ -56,7 +56,7 @@ Matcher::result_t Matcher::getSWResult(Sequence* dbSeq, const size_t seqDbSize,
             for (int32_t c = 0; c < alignment->cigarLen; ++c) {
                 char letter = SmithWaterman::cigar_int_to_op(alignment->cigar[c]);
                 uint32_t length = SmithWaterman::cigar_int_to_len(alignment->cigar[c]);
-                for (int i = 0; i < length; ++i){
+                for (uint32_t i = 0; i < length; ++i){
                     if (letter == 'M') {
                         if (dbSeq->int_sequence[targetPos] == currentQuery->int_sequence[queryPos]){
                             aaIds++;
